Move getByType to shared parsing utils
