+++ public/app/plugins/datasource/loki/querybuilder/parsing.ts
@@ -2,6 +2,7 @@ import { parser } from '@grafana/lezer-logql';
 import { SyntaxNode } from '@lezer/common';
 import {
   ErrorName,
+  getAllByType,
   getLeftMostChild,
   getString,
   makeBinOp,
@@ -423,27 +424,4 @@ function getLastChildWithSelector(node: SyntaxNode, selector: string) {
     }
   }
   return child;
-}
-
-/**
- * Get all nodes with type in the tree. This traverses the tree so it is safe only when you know there shouldn't be
- * too much nesting but you just want to skip some of the wrappers. For example getting function args this way would
- * not be safe is it would also find arguments of nested functions.
- * @param expr
- * @param cur
- * @param type
- */
-function getAllByType(expr: string, cur: SyntaxNode, type: string): string[] {
-  if (cur.name === type) {
-    return [getString(expr, cur)];
-  }
-  const values: string[] = [];
-  let pos = 0;
-  let child = cur.childAfter(pos);
-  while (child) {
-    values.push(...getAllByType(expr, child, type));
-    pos = child.to;
-    child = cur.childAfter(pos);
-  }
-  return values;
 }
